Fixed redirects for edit and delete

+++ src/main/java/travel/repository/VacationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import travel.beans.Vacation;
 
 @Repository
-public interface VacationRepository extends JpaRepository<Vacation, Long> {
+public interface VacationRepository extends JpaRepository<Vacation, Integer> {
 
 }
